Update 02-circlize-layout.Rmd
Line 630 (was 629) (image=) - This uses the numeric index of the sector to select the image source.  But the images in the circos have been ordered alphabetically.  So the images no longer correspond to the correct pokemon.  If you check the images in the example output the images do not match the names.
My code worked for me, but there may be a better way to extract the correct src for the image
new line 623 puts the name and src into a df ... and then I match the sector.index ...